feat: Print warning information when using a lower version of cryptogen, which does not affect use --story=1017123

+++ scripts/prepare_pk.sh
@@ -172,8 +172,12 @@ function generate_keys() {
 
 function generate_tls() {
 
-    cd "${BUILD_PATH}"
+    if [ ! -f "$CRYPTOGEN_TLS_CONF" ]; then
+        echo "WARN:File $CRYPTOGEN_TLS_CONF does not exist.Unable to generate tls certificate."
+        return 0
+    fi
 
+    cd "${BUILD_PATH}"
     cp $CRYPTOGEN_TLS_CONF crypto_tls_config.yml
     xsed "s%count: n%count: ${NODE_CNT}%g" crypto_tls_config.yml
